backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
apify-client==1.7.1
python-dotenv==1.0.0
requests==2.31.0
google-genai==1.0.0
openai==1.3.0
schedule==1.2.0